Adjustments to notifications graph endpoint
