+++ middleware/graphql/resolvers/query/notifications/NOTIFICATION_DISMISSABLE.js
@@ -1,0 +1,5 @@
+module.exports = {
+    NEVER: 0,
+    TEMPORARILY: 1,
+    PERMANENTLY: 2,
+}
