+++ middleware/graphql/resolvers/query/notifications/NOTIFICATION_SEVERITY.js
@@ -1,0 +1,5 @@
+module.exports = {
+    LOW: 0,
+    MEDIUM: 1,
+    HIGH: 2,
+}
