+++ middleware/graphql/schema.graphql
@@ -142,7 +142,7 @@ type Notification {
 	id: String
 	type: Int!
 	severity: Int!
-	dismissType: Int!
+	dismissable: Int!
 	text: String!
 	moreLink: String
 }
